feat: agregar hooks  para gestionar notificaciones y suscripciones MQTT
