+++ src/shared/hooks/notifications/index.ts
@@ -1,0 +1,9 @@
+// Exportar hooks de notificaciones
+export { useNotifications } from "../use-notifications";
+export { usePersonalNotifications } from "../use-personal-notifications";
+export { useGlobalNotifications } from "../use-global-notifications";
+
+// Re-exportar tipos para conveniencia
+export type { UseNotificationsReturn, UseNotificationsOptions } from "../use-notifications";
+export type { UsePersonalNotificationsReturn } from "../use-personal-notifications";
+export type { UseGlobalNotificationsReturn } from "../use-global-notifications";
